New RedShootBalls Autonomous implemented.
Might want to change something to guarantee we drive over the line at the
end, its cutting
it close (in my opinion).

Upon video review we believe that it's fine.

Also added drafts both a red and a blue gear autonomous that worked on the
practice field, but potentially
needs to be tuned with a drive back 2 inches command right after driving
forward.